requirements.txt
diffusers==0.30.0
Pillow==10.4.0
Requests==2.32.3
streamlit==1.37.1
torch==2.4.0+cu124
